// src/modules/players/application/use-cases/create-current-player-profile/create-current-player-profile.use-case.types.ts
import type { PlayerLevel } from '../../../domain/player-level';
import type { SportCode } from '../../../domain/sport-code';

type CreateCurrentPlayerProfileUseCaseRequest = {
  accessToken: string;
  displayName: string;
  mainSportCode: SportCode;
  mainSportLevel: PlayerLevel;
};

export type { CreateCurrentPlayerProfileUseCaseRequest };